finalize HSV integration + JPEG integration

+++ include/image_processing/segmentation.h
@@ -7,7 +7,7 @@
 #include "utils/matrix.h"
 
 namespace ip {
-std::shared_ptr<Matrix<uint8_t>> get_random_framing(std::shared_ptr<Matrix<uint8_t>> img, int n_tiles = 50);
+std::shared_ptr<Matrix<uint8_t>> get_random_framing(std::shared_ptr<Matrix<uint32_t>> img, int n_tiles = 50);
 }
 
 #endif
